Center main heading text
Using the body element as the container for the page this commit removes
any default margin applied to the body element to ensure consistency
accross browsers.

The h1 element is then targeted and the text-align property used to
center the text since heading elements are display: block by default so
the element already spans the full width of the page.