Change ul to nav element and center text
The un-ordered list element was not semantically accurate and
introduced an indent offset which I was unsure how to remove to ensure
the text was central.

This change wraps the links in div tags which in the future could be
used to create a more detailed summary card form of place-holder.

The nav element is suitable since it contains liks to all the pages of
the website.